refactor: rename tests utils to helpers
