+++ test/ImplDeployer.test.ts
@@ -4,8 +4,8 @@ import faker from "faker";
 import { ContractTransaction } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { expect } from ".";
-import { getTestUsers, TestUsers } from "./utils";
-import { computeInterfaceId } from "./utils/computeInterfaceId";
+import { getTestUsers, TestUsers } from "./helpers";
+import { computeInterfaceId } from "./helpers/computeInterfaceId";
 import { ContractInterfaces, AddressConstants } from "../src/common";
 import { deployImplDeployerFixture, deployTradeTrustERC721ImplFixture } from "./fixtures";
 import { encodeInitParams, getEventFromTransaction } from "../src/utils";
