+++ test/TitleEscrow.test.ts
@@ -6,8 +6,8 @@ import { Signer } from "ethers";
 import { FakeContract, MockContract, smock } from "@defi-wonderland/smock";
 import { expect } from ".";
 import { deployTokenFixture, deployTitleEscrowFixture } from "./fixtures";
-import { getTestUsers, getTitleEscrowContract, impersonateAccount, TestUsers } from "./utils";
-import { computeInterfaceId } from "./utils/computeInterfaceId";
+import { getTestUsers, getTitleEscrowContract, impersonateAccount, TestUsers } from "./helpers";
+import { computeInterfaceId } from "./helpers/computeInterfaceId";
 import { ContractInterfaces, AddressConstants } from "../src/common";
 import { deployImplProxy } from "./fixtures/deploy-impl-proxy.fixture";
